refactor: rename kuttl migrate to migrate kuttl
