+++ testdata/commands/migrate/kuttl/tests/help.txt
