+++ website/docs/commands/chainsaw.md
@@ -17,7 +17,7 @@ chainsaw [flags]
 * [chainsaw completion](chainsaw_completion.md)	 - Generate the autocompletion script for the specified shell
 * [chainsaw docs](chainsaw_docs.md)	 - Generate reference documentation
 * [chainsaw generate](chainsaw_generate.md)	 - Generate commands
-* [chainsaw kuttl](chainsaw_kuttl.md)	 - Work with KUTTL tests
+* [chainsaw migrate](chainsaw_migrate.md)	 - Migrate resources to Chainsaw
 * [chainsaw test](chainsaw_test.md)	 - Run tests
 * [chainsaw version](chainsaw_version.md)	 - Print the version informations
