--- website/docs/commands/chainsaw_kuttl.md
@@ -1,19 +1,0 @@
-## chainsaw kuttl
-
-Work with KUTTL tests
-
-```
-chainsaw kuttl [flags]
-```
-
-### Options
-
-```
-  -h, --help   help for kuttl
-```
-
-### SEE ALSO
-
-* [chainsaw](chainsaw.md)	 - Stronger tool for e2e testing
-* [chainsaw kuttl migrate](chainsaw_kuttl_migrate.md)	 - Migrate KUTTL resources to Chainsaw
-
