--- website/docs/commands/chainsaw_kuttl_migrate.md
@@ -1,20 +1,0 @@
-## chainsaw kuttl migrate
-
-Migrate KUTTL resources to Chainsaw
-
-```
-chainsaw kuttl migrate [flags]
-```
-
-### Options
-
-```
-  -h, --help   help for migrate
-```
-
-### SEE ALSO
-
-* [chainsaw kuttl](chainsaw_kuttl.md)	 - Work with KUTTL tests
-* [chainsaw kuttl migrate config](chainsaw_kuttl_migrate_config.md)	 - Migrate KUTTL config to Chainsaw
-* [chainsaw kuttl migrate tests](chainsaw_kuttl_migrate_tests.md)	 - Migrate KUTTL tests to Chainsaw
-
