+++ website/docs/commands/chainsaw_migrate.md
@@ -1,0 +1,19 @@
+## chainsaw migrate
+
+Migrate resources to Chainsaw
+
+```
+chainsaw migrate [flags]
+```
+
+### Options
+
+```
+  -h, --help   help for migrate
+```
+
+### SEE ALSO
+
+* [chainsaw](chainsaw.md)	 - Stronger tool for e2e testing
+* [chainsaw migrate kuttl](chainsaw_migrate_kuttl.md)	 - Migrate KUTTL resources to Chainsaw
+
