+++ website/docs/commands/chainsaw_migrate_kuttl.md
@@ -1,0 +1,20 @@
+## chainsaw migrate kuttl
+
+Migrate KUTTL resources to Chainsaw
+
+```
+chainsaw migrate kuttl [flags]
+```
+
+### Options
+
+```
+  -h, --help   help for kuttl
+```
+
+### SEE ALSO
+
+* [chainsaw migrate](chainsaw_migrate.md)	 - Migrate resources to Chainsaw
+* [chainsaw migrate kuttl config](chainsaw_migrate_kuttl_config.md)	 - Migrate KUTTL config to Chainsaw
+* [chainsaw migrate kuttl tests](chainsaw_migrate_kuttl_tests.md)	 - Migrate KUTTL tests to Chainsaw
+
