+++ website/docs/more/kuttl-migration.md
@@ -2,19 +2,19 @@
 
 ## Overview
 
-The `chainsaw kuttl migrate tests` and `chainsaw kuttl migrate config` commands are designed for the migration of KUTTL tests to Chainsaw.
+The `chainsaw migrate kuttl tests` and `chainsaw migrate kuttl config` commands are designed for the migration of KUTTL tests to Chainsaw.
 
-- [`chainsaw kuttl migrate config`](#migrate-configuration)
+- [`chainsaw migrate kuttl config`](#migrate-configuration)
 
     migrates a KUTTL `TestSuite` to the corresponding Chainsaw `Configuration`
 
-- [`chainsaw kuttl migrate tests`](#migrate-tests)
+- [`chainsaw migrate kuttl tests`](#migrate-tests)
 
     migrates KUTTL tests to the corresponding Chainsaw `Test`s
 
 !!! tip "Reference documentation"
 
-    You can view the full command documentation [here](../commands/chainsaw_kuttl_migrate.md).
+    You can view the full command documentation [here](../commands/chainsaw_migrate_kuttl.md).
 
 ## Examples
 
@@ -23,7 +23,7 @@ The `chainsaw kuttl migrate tests` and `chainsaw kuttl migrate config` commands 
 The command below will migrate KUTTL tests to Chainsaw and overwrite original files with converted ones.
 
 ```bash
-chainsaw kuttl migrate tests path/to/kuttl/tests --save --cleanup
+chainsaw migrate kuttl tests path/to/kuttl/tests --save --cleanup
 ```
 
 This will generate a `chainsaw-test.yaml` for every KUTTL test discovered.
@@ -33,7 +33,7 @@ This will generate a `chainsaw-test.yaml` for every KUTTL test discovered.
 The command below will migrate a KUTTL test suite file to the corresponding Chainsaw `Configuration`.
 
 ```bash
-chainsaw kuttl migrate config path/to/kuttl/testsuite --save --cleanup
+chainsaw migrate kuttl config path/to/kuttl/testsuite --save --cleanup
 ```
 
 This will generate a `.chainsaw.yaml` configuration file.
